+++ packages/odyssey-react-mui/src/labs/index.ts
@@ -17,6 +17,8 @@ export type { LocalizationProviderProps } from "@mui/x-date-pickers";
 export * from "./DatePicker";
 export * from "./datePickerTheme";
 
+export * from "./DataComponents";
+
 /** @deprecated Will be removed in a future Odyssey version in lieu of the one shipping with DataTable */
 export * from "./DataTablePagination";
 export * from "./DataFilters";
@@ -32,3 +34,6 @@ export * from "./PaginatedTable";
 export * from "./GroupPicker";
 
 export * from "./Switch";
+
+export * from "./NavAccordion";
+export * from "./SideNav";